+++ Codec/Compression/GZip.hs
@@ -44,7 +44,7 @@ module Codec.Compression.GZip (
   DecompressParams(..), defaultDecompressParams,
 
   -- ** The compression parameter types
-  CompressionLevel,
+  CompressionLevel(..),
     defaultCompression,
     noCompression,
     bestSpeed,